Add Payee endpoints and a editorconfig
This commit adds in the two payee endpoints for the YNAB api, as well
as adding a test in the included console program. To ensure consitency
a .editorconfig file was added in to comply with the author's coding
style. Added in accessToken.txt into the .gitignore for easier local
testing.